ODROID-XU3/4: r8152: Added force device power reset
The ODROID-XU3/4 ethernet device is dropped with the -EPROTO status on
some specific network environment. The state is smilar to unplugging the
USB device or hub. But, ODROID-XU3/4 does not disconnect the eternet USB
device or hub. So, the ethernet device which is connected by the usb bus
interface could not be reset.

This change triggers the power of the r8153 ethernet device in ODROID-XU3/4
for the unexpected USB request block status(a.k.a> -EPROTO).

Change-Id: I465936131d8ac7f39e155b96238d3f98bad0eda1
Signed-off-by: Brian Kim <brian.kim@hardkernel.com>