ODROID-XU3/4 : Adjust hdmi timing for 800x480p60hz to support Vu5A
This commit is to support Vu5A with a higher frame rate
to improve display quality for emulation and gaming.

- Vu5A modeline information
Detailed mode: Clock 33.900 MHz, 476 mm x 268 mm
               800  844  932 1056 hborder 0
               480  483  489  535 vborder 0
              +hsync +vsync

To support Vu5A, pixel clock 33.9Mhz is needed,
but, we don't have the exact hdmi phy table,
so, as a workaround, the closest table, 36MHz will be used.

- ideal pixel clock : 33.9MHz / vsync 60hz
- alternate pixel clock : 36MHz / vsync 63.7Hz

Change-Id: I3a1e8da881ed9b433c29bcee341f0094385b72c5